requirements_old.txt
pandas
gym
numpy <
dill
matplotlib
torch
tqdm
gitpython
scikit_learn
tensorboardX
tables
#mujoco-py==1.50.1.68
packaging>=20.0
pytest>=5.0
appdirs>=1.4
oslo.concurrency>=4.2
